refactored everything else
