+++ app/mongo/mongo_bayes_helper.py
