+++ app/routes/views.py
@@ -1,9 +1,9 @@
 from app import shenanigans
 from flask import render_template, request
-from pages import *
+from app.pages import *
 #from mongo_helper import *
 from API import parser
-import mongo_helper as mh
+import app.mongo.mongo_helper as mh
 
 # Adding form components here
 #from components import *
