+++ app/util/data/filters.py
@@ -4,7 +4,7 @@ import re
 import unicodedata
 import sys
 import os
-from constants import *
+from app.common.constants import *
 from app.common.file_paths import APP_ROOT
 from filter_helper import *
 from nltk.stem import WordNetLemmatizer
