+++ app/util/data/recipe_parser.py
@@ -1,5 +1,5 @@
 from filters import *
-import mongo_helper as mh
+import app.mongo.mongo_helper as mh
 import collections as c
 import pickle
 import os
